requirements.txt
torch>=2.2.0
fire==0.5.0
pyarrow==15.0.0
transformers==4.40.2
ibm-fms>=0.0.3
